[FIX] 패키지 위치 변경 및 주석 추가
