+++ src/main/java/com/example/market/security/config/SecurityConfig.java
@@ -1,8 +1,12 @@
 package com.example.market.security.config;
 
-import com.example.market.security.handler.LoginFailureHandler;
-import com.example.market.security.handler.LoginSuccessHandler;
-import com.example.market.security.jwt.*;
+import com.example.market.security.handler.jwt.JwtAccessDeniedHandler;
+import com.example.market.security.handler.jwt.JwtAuthenticationEntryPoint;
+import com.example.market.security.handler.login.LoginFailureHandler;
+import com.example.market.security.handler.login.LoginSuccessHandler;
+import com.example.market.security.jwt.JwtAuthenticationFilter;
+import com.example.market.security.jwt.LoginProvider;
+import com.example.market.security.jwt.TokenProvider;
 import lombok.RequiredArgsConstructor;
 import org.springframework.boot.autoconfigure.security.servlet.PathRequest;
 import org.springframework.context.annotation.Bean;
