--- src/main/java/com/example/market/security/handler/JwtAccessDeniedHandler.java
@@ -1,22 +1,0 @@
-package com.example.market.security.handler;
-
-import org.springframework.security.web.access.AccessDeniedHandler;
-import org.springframework.stereotype.Component;
-
-import javax.servlet.ServletException;
-import javax.servlet.http.HttpServletRequest;
-import javax.servlet.http.HttpServletResponse;
-import java.io.IOException;
-import org.springframework.security.access.AccessDeniedException;
-
-@Component
-public class JwtAccessDeniedHandler implements AccessDeniedHandler {
-    //필요한 권한이 없이 접근하려 할때 403
-
-    @Override
-    public void handle(HttpServletRequest request, HttpServletResponse response,
-                       AccessDeniedException accessDeniedException) throws IOException, ServletException {
-        response.sendError(HttpServletResponse.SC_FORBIDDEN);
-
-    }
-}
